+++ leaspy/models/utils/attributes/__init__.py
@@ -1,4 +1,4 @@
-from .attributes_logistic import AttributesLogistic
-from .attributes_linear import AttributesLinear
-from .attributes_logistic_parallel import AttributesLogisticParallel
+from .logistic_attributes import LogisticAttributes
+from .linear_attributes import LinearAttributes
+from .logistic_parallel_attributes import LogisticParallelAttributes
 from .attributes_factory import AttributesFactory