+++ tests/unit_tests/test_leaspy_import.py
@@ -15,6 +15,12 @@ class ImportTest(unittest.TestCase):
             print(lp)
         self.__import_or_fail(__import)
 
+    def test_import_main_klass_from(self):
+        def __import():
+            from leaspy import Leaspy, Data, AlgorithmSettings, IndividualParameters, __watermark__
+            print(__watermark__)
+        self.__import_or_fail(__import)
+
     def __import_or_fail(self, import_function):
         try:
             import_function()